requirements.txt
asgiref==3.5.0
certifi==2021.10.8
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.2
django-adminlte2==0.4.3
django-stdimage==5.3.0
gunicorn==20.1.0
Pillow==9.0.1
progressbar2==4.0.0
psycopg2-binary==2.9.3
python-utils==3.1.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
static3==0.7.0
